+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add default "help" sub-command to all commands.
 - Allow for custom command lookups by overriding SubOptParser#[].
+- Allow for adding commands are parse time with .on_parse { |so| ... }
 
 ## [0.1.0] - 2025-01-21